Config mailcatcher
according to:
 - https://mailcatcher.me
 - https://github.com/dockage/mailcatcher/blob/main/Dockerfile
 - https://courtneyzhan.medium.com/mailcatcher-in-docker-c2ef1f945a0b
 - https://stackoverflow.com/questions/61630292/error-connection-refused-mailcatcher-in-docker-rails
There were some issues with the connection, however
after searching trough the posts the final output seems to work.
However I am not sure about the correctness of this approach,
especially when it comes to the docker files.

It is important to notice that in the docker-compose
the new profile is created as 'mailcatcher' in order
to make it work in the corrensponsing env file
the profile needs to be exported [1].

[1]: https://github.com/ciupkajakub/shop/blob/main/.env.example#L21