Add Cart logic
based on https://www.youtube.com/watch?v=SPokmOwiM7E.
However it is done only for the part of this tutorial.

It is basic mechanism of the shopping cart.
However what I would like to do next is
to get rid of intializing the cart when it is not
needed. I would create it only when someone adds
something to the cart. And when someone visits
the /cart url, just render that is empty. We do
not need empty cart in the DB.

The next step is to create a User and implement
the logic for rewriting the shopping cart from
a guest to a user who logs in after adding some
products.